requirements.txt
#devops
black>=25.1
click>=8.2
pytest>=8.4
pytest-cov>=6.2
pylint>=3.3
#data
jupyter>=1.1
pandas>=2.3
numpy>=2.3
pydantic>=2.11
#datascience
scikit-learn>=1.7
matplotlib>=3.10
seaborn>=0.13
dask>=2025.7
#geo
geopy>=2.4
rasterio>=1.4
shapely>=2.1
geopandas>=1.1
pyarrow>=21.0
fiona>=1.10
pyshp>=2.3
#api
fastapi>=0.116
uvicorn>=0.35
python-multipart>=0.0.20
aiofiles>=24.1
fastapi-pagination>=0.13
httpx>=0.23
pyjwt>=2.10
passlib[bcrypt]>=1.7
pyjwt>=2.10
python-jose[cryptography]>=3.5
cachetools>=6.1
pydantic-settings>=2.10
authlib>=1.6
python-dotenv>=1.1
itsdangerous>=2.2